week_1. Реализовал создание заказа

+++ shared/pkg/openapi/order/v1/oas_json_gen.go
@@ -490,7 +490,7 @@ func (s *CreateOrderResponse) encodeFields(e *jx.Encoder) {
 	}
 	{
 		e.FieldStart("total_price")
-		e.Float32(s.TotalPrice)
+		e.Float64(s.TotalPrice)
 	}
 }
 
@@ -523,8 +523,8 @@ func (s *CreateOrderResponse) Decode(d *jx.Decoder) error {
 		case "total_price":
 			requiredBitSet[0] |= 1 << 1
 			if err := func() error {
-				v, err := d.Float32()
-				s.TotalPrice = float32(v)
+				v, err := d.Float64()
+				s.TotalPrice = float64(v)
 				if err != nil {
 					return err
 				}
@@ -991,7 +991,7 @@ func (s *OrderDto) encodeFields(e *jx.Encoder) {
 	}
 	{
 		e.FieldStart("total_price")
-		e.Float32(s.TotalPrice)
+		e.Float64(s.TotalPrice)
 	}
 	{
 		e.FieldStart("transaction_uuid")
@@ -1073,8 +1073,8 @@ func (s *OrderDto) Decode(d *jx.Decoder) error {
 		case "total_price":
 			requiredBitSet[0] |= 1 << 3
 			if err := func() error {
-				v, err := d.Float32()
-				s.TotalPrice = float32(v)
+				v, err := d.Float64()
+				s.TotalPrice = float64(v)
 				if err != nil {
 					return err
 				}
