+++ shared/pkg/openapi/order/v1/oas_schemas_gen.go
@@ -133,7 +133,7 @@ type CreateOrderResponse struct {
 	// Идентификатор заказа.
 	OrderUUID string `json:"order_uuid"`
 	// Общая сумма заказа.
-	TotalPrice float32 `json:"total_price"`
+	TotalPrice float64 `json:"total_price"`
 }
 
 // GetOrderUUID returns the value of OrderUUID.
@@ -142,7 +142,7 @@ func (s *CreateOrderResponse) GetOrderUUID() string {
 }
 
 // GetTotalPrice returns the value of TotalPrice.
-func (s *CreateOrderResponse) GetTotalPrice() float32 {
+func (s *CreateOrderResponse) GetTotalPrice() float64 {
 	return s.TotalPrice
 }
 
@@ -152,7 +152,7 @@ func (s *CreateOrderResponse) SetOrderUUID(val string) {
 }
 
 // SetTotalPrice sets the value of TotalPrice.
-func (s *CreateOrderResponse) SetTotalPrice(val float32) {
+func (s *CreateOrderResponse) SetTotalPrice(val float64) {
 	s.TotalPrice = val
 }
 
@@ -378,7 +378,7 @@ type OrderDto struct {
 	// Список идентификаторов запчастей.
 	PartUuids []string `json:"part_uuids"`
 	// Сумма заказа.
-	TotalPrice float32 `json:"total_price"`
+	TotalPrice float64 `json:"total_price"`
 	// Идентификатор транзакции.
 	TransactionUUID string        `json:"transaction_uuid"`
 	PaymentMethod   PaymentMethod `json:"payment_method"`
@@ -401,7 +401,7 @@ func (s *OrderDto) GetPartUuids() []string {
 }
 
 // GetTotalPrice returns the value of TotalPrice.
-func (s *OrderDto) GetTotalPrice() float32 {
+func (s *OrderDto) GetTotalPrice() float64 {
 	return s.TotalPrice
 }
 
@@ -436,7 +436,7 @@ func (s *OrderDto) SetPartUuids(val []string) {
 }
 
 // SetTotalPrice sets the value of TotalPrice.
-func (s *OrderDto) SetTotalPrice(val float32) {
+func (s *OrderDto) SetTotalPrice(val float64) {
 	s.TotalPrice = val
 }
